refactor: simplify context and services definition
