+++ src/app/layout.tsx
@@ -1,8 +1,6 @@
 'use client'
 
 import React from 'react'
-import { TdpClientContextProvider } from 'src/contexts'
-
 import 'src/styles/globals.css'
 
 export default function RootLayout({
@@ -12,9 +10,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>
-        <TdpClientContextProvider>{children}</TdpClientContextProvider>
-      </body>
+      <body>{children}</body>
     </html>
   )
 }
